Refactor nested classes into their own files
